fixing the binary trie stuff
