+++ hgeometry-combinatorial/hgeometry-combinatorial.cabal
@@ -149,16 +149,24 @@ library
         HGeometry.Sign
         HGeometry.StringSearch.KMP
 
+        -- HGeometry.Small.ZeroOrTwo
+        HGeometry.Small.AtMostTwo
+
         HGeometry.Tree.Binary.Static
         HGeometry.Tree.Util
+
+
         HGeometry.Trie
+        HGeometry.Trie.Binary
 
         HGeometry.Unbounded
         HGeometry.Vector.NonEmpty.Util
 
         HGeometry.YAML
 
-    -- other-modules:
+    other-modules:
+        HGeometry.Trie.Type
+
     -- other-extensions:
 
     hs-source-dirs:   src
